fix verbose output, prepare compat.encode()
On Python 2 a byte string was actually encoded to a byte string again
when the VERBOSE flag was active.
This was fixed and at the same time made Python 3 compliant
as in the previous commit.

compat.encode() was also added for later use.
After the double encode fix it is currently not used.